* synchronizing changes to traffic simulator * synchronizing changes to configuration files * synchronizing changes to inlined functions
git-svn-id: http://madara.googlecode.com/svn/trunk@198 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ madara/knowledge_engine/Knowledge_Base.h
@@ -62,7 +62,7 @@ namespace Madara
       void close_transport (void);
 
       /// Retrieve the value associated with the provided key
-      int get (const ::std::string & key) const;
+      long get (const ::std::string & key) const;
 
       /// Expand a statement by variable expansions 
       std::string expand_statement (const ::std::string & statement) const;
